==> scripts/1_nasmar_cartografia.ipynb <==
# Situación geográfica del problema

En este `jupyter` mostramos algunos mapas de Honduras y los municipios de NASMAR para:

* Analizar la Cartografía Base que está introducida en las bases de datos de gvSIG-Fonsagua (los ficheros `fonsagua.sqlite`)
* Ubicar la zona de trabajo
* Generar algunas variables, funciones, etc... que se puedan emplear en el resto de los análisis y fases del trabajo

_Nota: Cuando hacemos referencia al INE es al censo oficial (del 2002) del Instituto Nacional de Estadística de Honduras_

En la base de datos de Fonsagua tenemos las siguientes capas de cartografía base de límites administrativos:

* `departamentos`. Con todos los departamentos de Honduras
* `municipios`. Son todos los de Valle y Choluteca
* `aldeas`. Son todas las de Valle y Choluteca. Los códigos y nombres son del INE
* `caserios`. Son todos los del INE los pertenecientes a 9 municipios. Los 7 de NASMAR + Choluteca ('0601') + Alianza ('1702').


La zona de trabajo de NASMAR son 7 municipios de (de un total de 25) pertenecientes a los departamentos de Valle ('17') y Choluteca ('06'):

* '0606', 'El Triunfo'
* '0607', 'Marcovia'
* '0609', 'Namasigue'
* '1701', 'Nacaome'
* '1703', 'Amapala'
* '1706', 'Goascaran'
* '1709', 'San Lorenzo'

# Configuración básica de jupyter, pandas, ...
# e importación de librerías que necesitaremos durante el análisis

# https://medium.com/@1522933668924/using-matplotlib-in-jupyter-notebooks-comparing-methods-and-some-tips-python-c38e85b40ba1
# Para hacer los mapas interactivos

# %matplotlib widget 
%matplotlib inline 
import numpy as np
from IPython.display import display, HTML, Markdown
# Ejemplos de uso
# display(HTML('<h2 style="color:red;">Subtitulo "YYY - NOMBRE"</h2>'))
# display(Markdown("# Título principal"))
# display(HTML(df.describe().to_html()))  # muestra todos los datos
import pandas
from ietl import pandas_utils

from ietl.geopandas_utils import (
    get_dataframe_from_spatialite_table,
    plot_polygon_labels,
)
import geopandas as gpd

pandas.options.mode.use_inf_as_na = True
# http://songhuiming.github.io/pages/2017/04/02/jupyter-and-pandas-display/
# pandas.set_option('display.max_rows', 500)
# pandas.set_option('display.max_columns', 500)


# https://ipython.readthedocs.io/en/stable/config/extensions/autoreload.html
# Si se está trabajando con módulos externos usar %autoreload 2 evita tener que
# recargar a mano. %autoreload 0 lo desactiva
%load_ext autoreload
%autoreload 2



# Ejecuta el .ipyndb referenciando como si fuera en este propio Lab
# Todas las variables del otro pasan a estar disponible en este
# Cualquier salida por pantalla del otro se muestra en los resultados 
# de la celda donde se ejecuta el %run
# if __name__ == '__main__' and '__file__' not in globals():
#     print("Este código sólo se ejecuta cuando lo llamo directamente y no a través de un `%run`")
# %run nasmar_cartografia.ipynb

from nasmar import common_left_field, common_right_field
import nasmar

sqlite_path = "191208_fonsagua_todos_municipios.sqlite"
excel_datos_pdm_path = "pdm.xlsx"

# DEPARTAMENTOS

departamentos_plot_options = {
    "edgecolor": "lightgrey", # "black"
    "color": "whitesmoke", #None,
    "alpha": 0.5,
    "linewidth": 0.5, # None
}
departamentos = nasmar.get_departamentos_bd()
ax = departamentos.plot(figsize=(10, 10), **departamentos_plot_options)
ax.set_axis_off()

plot_polygon_labels(departamentos, "nombre", {"fontsize": 8})
# display(departamentos[["codigo", "nombre"]])

# MUNICIPIOS

municipios = nasmar.get_municipios_bd()
nasmar.plot_municipios(municipios)

# display(municipios[["codigo", "nombre"]])
# municipios.to_file("municipios.shp", encoding = 'utf-8')


# ALDEAS

# En bd llamamos a esta tabla "cantones" por la terminología de El Salvador.
aldeas = nasmar.get_aldeas_bd()

aldeas_plot_options = {
    "edgecolor": "lightgrey", # "black"
    # "color": "whitesmoke", #None,
    "alpha": 0.5,
    "linewidth": 0.5, # None
    "column": "nombre",
}
aldeas_ax = aldeas.plot(figsize=(10, 10), **aldeas_plot_options)
aldeas_ax.set_axis_off()
plot_polygon_labels(aldeas, "nombre", {"fontsize": 8})
# display(aldeas[["codigo", "nombre"]])

# CASERIOS
# En bd llamamos a esta tabla "caserios_comunidades"

caserios = nasmar.get_caserios_bd()

# Contamos el número de elementos del df que tenemos por municipio
display(nasmar.number_of_elements_by_council(caserios, "codigo"))
# Eliminamos los caserios que no están en los municipios NASMAR
nasmar.drop_elements_not_in_nasmar_councils(caserios, "codigo")
# Volvemos a hacer el conteo    
display(nasmar.number_of_elements_by_council(caserios, "codigo"))


caserios_plot_options = {
    "edgecolor": "lightgrey", # "black"
    # "color": "whitesmoke", #None,
    "alpha": 0.5,
    "linewidth": 0.5, # None
    # "column": "nombre",
}
caserios_ax = caserios.plot(figsize=(10, 10), **caserios_plot_options)
caserios_ax.set_axis_off()
# plot_polygon_labels(caserios, "nombre", {"fontsize": 8})
display(caserios)
# caserios.to_file("caserios.shp", encoding = 'utf-8')
